requirements.txt
transformers
torch
numind
pdfplumber
litellm
streamlit
python-dotenv
scikit-learn
pandas
sentence-transformers